# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ALMEAA API
  version: 0.4.0
paths:
...
  /api/v1/auth/admin/users:
    get:
      operationId: adminListUsers
      parameters:
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { in: query, name: search, schema: { type: string, maxLength: 120 } }
        - { in: query, name: role, schema: { type: string, enum: [student, teacher, admin, supervisor, school_admin, parent] } }
        - { in: query, name: isActive, schema: { type: boolean } }
        - { in: query, name: platformTrainer, schema: { type: boolean } }
      responses:
        "200": { description: Paginated user directory; supervisors and teachers are limited to shared active school/class scope }
        "401": { description: Authentication required }
        "403": { description: Role not permitted }
    post:
      operationId: adminUpsertUser
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, email, password, role]
              properties:
                name: { type: string, minLength: 2 }
                email: { type: string, format: email }
                password: { type: string, minLength: 8, maxLength: 160 }
                role: { type: string, enum: [student, teacher, admin, supervisor, school_admin, parent] }
                schoolId: { type: [string, "null"] }
                groupIds: { type: array, items: { type: string } }
                linkedStudentIds: { type: array, items: { type: string } }
                managedPathIds: { type: array, maxItems: 50, items: { type: string }, description: Compatibility form mapped transactionally to canonical Content trainer scope; teacher role only. }
                managedSubjectIds: { type: array, maxItems: 200, items: { type: string }, description: Compatibility form mapped transactionally to canonical Content trainer scope; teacher role only. }
      responses:
        "201": { description: User account created or updated by email }
        "400": { description: Invalid request or last-admin invariant }
        "401": { description: Authentication required }
        "403": { description: Admin role required }
        "409": { description: Managed trainer scope conflicts with role or active taxonomy }
  /api/v1/auth/admin/users/summary:
    get:
      operationId: adminUsersSummary
      responses:
        "200": { description: User totals, inactive count, counts by role, and canonical active platform-trainer count }
        "401": { description: Authentication required }
        "403": { description: Admin role required }
...
  /api/v1/auth/admin/users/{id}:
    patch:
      operationId: adminUpdateUser
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: { type: string, minLength: 2, maxLength: 120 }
                avatar: { type: string, maxLength: 2000 }
                role: { type: string, enum: [student, teacher, admin, supervisor, school_admin, parent] }
                isActive: { type: boolean }
                schoolId: { type: [string, "null"] }
                groupIds: { type: array, items: { type: string } }
                linkedStudentIds: { type: array, items: { type: string } }
                managedPathIds: { type: array, maxItems: 50, items: { type: string }, description: Compatibility form mapped transactionally to canonical Content trainer scope; teacher role only. Empty array revokes path scope. }
                managedSubjectIds: { type: array, maxItems: 200, items: { type: string }, description: Compatibility form mapped transactionally to canonical Content trainer scope; teacher role only. Empty array revokes subject scope. }
      responses:
        "200": { description: User account updated }
        "400": { description: Invalid request or operation would remove last active admin }
        "401": { description: Authentication required }
        "403": { description: Admin role required }
        "404": { description: User not found }
        "409": { description: Organization or managed trainer scope conflicts with canonical relations }
    delete:
      operationId: adminDeleteUser
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string }
      responses:
        "200": { description: User deleted }
        "400": { description: Self-delete or last-admin deletion blocked }
        "401": { description: Authentication required }
        "403": { description: Admin role required }
        "404": { description: User not found }
...